clean up content
