--- src/main/steps/applicant1/hub-page/pages/sole/hub-1.njk
@@ -1,12 +1,0 @@
-{% if progressionIndex == 0 and not hasApplicantConfirmedReceipt %}
-  {% for property, line in applicationSubmittedLatestUpdate %}
-    <p class="govuk-body">{{ line }}</p>
-  {% endfor %}
-  {% block form %}
-    {% include "../../form.njk" %}
-  {% endblock %}
-{% else %}
-  {% for property, line in applicantConfirmedReceiptLatestUpdate %}
-    <p class="govuk-body">{{ line }}</p>
-  {% endfor %}
-{% endif %}
