+++ src/main/steps/applicant1/hub-page/sole/hub-1.njk
@@ -1,0 +1,3 @@
+<p class="govuk-body">{{ subLine1 }}</p>
+<p class="govuk-body">{{ subLine2 }}</p>
+<p class="govuk-body">{{ subLine3 }}</p>
