+++ src/main/steps/applicant1/hub-page/template.njk
@@ -55,9 +55,9 @@
       <h2 class="govuk-heading-m">{{ subHeading1 }}</h2>
       <div class="govuk-panel govuk-panel--outstanding">
         {% if isJointApplication %}
-          {% include "./pages/joint/hub-1.njk" %}
+          {% include "./joint/hub-1.njk" %}
         {% else %}
-          {% include "./pages/sole/hub-1.njk" %}
+          {% include "./sole/hub-1.njk" %}
         {% endif %}
 
         <h3 class="govuk-heading-s">{{ subHeading2 }}</h3>
